+++ chevah/compat/interfaces.py
@@ -197,9 +197,12 @@ class IOSUsers(Interface):
         switching to user.
         """
 
-    def isUserInGroups(username, groups, token):
+    def getGroupForUser(username, groups, token):
         """
-        Return true if `username` or 'token' is a member of `groups`.
+        Return the name of the first group from `groups` that `username` is
+        member of.
+
+        Retrun None if `username` is not a member of any of the groups.
         """
 
     def getPrimaryGroup(username):